rename module.rb and class.rb to model_as_module.rb and model_as_class.rb
