+++ lib/metaruby.rb
@@ -3,8 +3,8 @@ require 'weakref'
 
 require 'metaruby/inherited_attribute'
 require 'metaruby/registration'
-require 'metaruby/module'
-require 'metaruby/class'
+require 'metaruby/model_as_module'
+require 'metaruby/model_as_class'
 
 require 'utilrb/logger'
