+++ lib/metaruby/model_as_class.rb
