+++ lib/metaruby/model_as_module.rb
