+++ test/test_model_as_class.rb
