+++ test/test_model_as_module.rb
